requirements.txt
PyQt6
requests
httpx
mysql-connector-python
psycopg2-binary
google-generativeai